+++ core/tests.py
@@ -1,0 +1,46 @@
+from django.test import TestCase, SimpleTestCase
+from django.urls import reverse
+from unittest import mock
+from core.sitemaps import StaticViewSitemap
+from core.redis_client import RedisClient
+
+
+class HomepageViewTests(TestCase):
+    def test_homepage_status_code(self):
+        response = self.client.get(reverse('core:core_home'))
+        self.assertEqual(response.status_code, 200)
+
+    def test_homepage_template_used(self):
+        response = self.client.get(reverse('core:core_home'))
+        self.assertTemplateUsed(response, 'core/index.html')
+
+
+class StaticViewSitemapTests(SimpleTestCase):
+    def setUp(self):
+        self.sitemap = StaticViewSitemap()
+
+    def test_items(self):
+        self.assertEqual(self.sitemap.items(), ['core:homepage'])
+
+    def test_location(self):
+        self.assertEqual(self.sitemap.location('core:homepage'), '/')
+
+
+class RedisClientTests(TestCase):
+    @mock.patch('core.redis_client.settings')
+    @mock.patch('core.redis_client.Redis')
+    def test_get_instance_creates_singleton(self, mock_redis_class, mock_settings):
+        mock_settings.REDIS_HOST = 'localhost'
+        mock_settings.REDIS_PORT = 6379
+        mock_settings.REDIS_DB = 0
+
+        # Reset instance for test isolation
+        RedisClient._instance = None
+
+        instance1 = RedisClient.get_instance()
+        instance2 = RedisClient.get_instance()
+
+        self.assertIs(instance1, instance2)
+        mock_redis_class.assert_called_once_with(
+            host='localhost', port=6379, db=0, decode_responses=True
+        )
